move language attribute test to validation tests
