+++ org.lflang.tests/src/org/lflang/tests/compiler/LinguaFrancaValidationTest.java
@@ -2285,6 +2285,28 @@ public class LinguaFrancaValidationTest {
     }
 
     @Test
+    public void testLanguageAttribute() throws Exception {
+        String testCase = """
+                target Python;
+                main reactor {
+                    @language(value = "C")
+                    reaction(startup) {= =}
+                }
+            """;
+        validator.assertNoIssues(parseWithoutError(testCase));
+
+        testCase = """
+                target Python;
+                main reactor {
+                    @language(value = "Foo")
+                    reaction(startup) {= =}
+                }
+            """;
+        validator.assertError(parseWithoutError(testCase), LfPackage.eINSTANCE.getAttribute(), null,
+                               "Incorrect language: value should be a supported language");
+    }
+
+    @Test
     public void testInitialMode() throws Exception {
         String testCase = """
             target C;
